automatically backed up by learn

+++ index.js
@@ -28,7 +28,7 @@ function makeInt(n){
 }
 
 function preserveDecimal(n){
-  return parseInt(n);
+  return parseFloat(n);
 }
